requirements.txt
aiohttp==3.13.5
boto3==1.42.91
davey==0.1.5
discord.py==2.7.1
discord-ext-menus==1.1
fuzzywuzzy==0.18.0
matplotlib==3.10.8
numpy==2.4.4
openai==2.32.0
opencv-python==4.13.0.92
Pillow==12.2.0
PyNaCl==1.6.2
pyfiglet==1.0.4
pytz==2026.1.post1
requests==2.33.1
scipy==1.17.1
seaborn==0.13.2
pytest==9.0.3
pytest-asyncio==1.3.0
